requirements.txt
crewai
crewai-tools
fastapi
uvicorn[standard]
python-dotenv
litellm
sqlalchemy
psycopg2-binary
